+++ tests/testthat/test_rdf_to_tbl.R
@@ -3,7 +3,7 @@ library(RWDataPlyr)
 context("check rdf_to_rwtbl()")
 
 # check the default call -----------------
-rdftbl <- rdf_to_rwtbl(keyRdf)
+rdftbl <- expect_warning(rdf_to_rwtbl(keyRdf))
 reqCols <- RWDataPlyr:::req_rwtbl_cols()
 exp_atts <- c("mrm_config_name", "owner", "description", "create_date", 
               "n_traces")
@@ -42,7 +42,7 @@ test_that("no values are modified", {
 })
 
 # check the add_ym options ---------------
-rdftbl2 <- rdf_to_rwtbl(keyRdf, add_ym = FALSE)
+rdftbl2 <- expect_warning(rdf_to_rwtbl(keyRdf, add_ym = FALSE))
 test_that("dimensions are as expected with no ym", {
   expect_true(nrow(rdftbl2) > 0)
   expect_equal(ncol(rdftbl2), length(reqCols))
@@ -54,15 +54,17 @@ test_that("no ym tbl matches orig tbl", {
 })
 
 test_that("invalid add_ym values cause errors", {
-  expect_error(rdf_to_rwtbl(keyRdf, add_ym = NA))
-  expect_error(rdf_to_rwtbl(keyRdf, add_ym = "true"))
-  expect_error(rdf_to_rwtbl(keyRdf, add_ym = 7))
-  expect_error(rdf_to_rwtbl(keyRdf, add_ym = c(FALSE, TRUE, NA)))
+  expect_warning(expect_error(rdf_to_rwtbl(keyRdf, add_ym = NA)))
+  expect_warning(expect_error(rdf_to_rwtbl(keyRdf, add_ym = "true")))
+  expect_warning(expect_error(rdf_to_rwtbl(keyRdf, add_ym = 7)))
+  expect_warning(
+    expect_error(rdf_to_rwtbl(keyRdf, add_ym = c(FALSE, TRUE, NA)))
+  )
 })
 
 # check the scenario option -------------
-rdftbl2 <- rdf_to_rwtbl(keyRdf, scenario = "DNF,CT")
-rdftbl3 <- rdf_to_rwtbl(keyRdf, scenario = 1, add_ym = FALSE)
+rdftbl2 <- expect_warning(rdf_to_rwtbl(keyRdf, scenario = "DNF,CT"))
+rdftbl3 <- expect_warning(rdf_to_rwtbl(keyRdf, scenario = 1, add_ym = FALSE))
 test_that("dimensions are as expected with no ym", {
   expect_true(nrow(rdftbl2) > 0)
   expect_equal(ncol(rdftbl2), length(reqCols) + 3)
@@ -87,18 +89,20 @@ test_that("no add scenario tbl matches orig tbl", {
 })
 
 test_that("scenario options error properly", {
-  expect_error(rdf_to_rwtbl(keyRdf, scenario = c("DNF,CT", "DNF,C1")))
-  expect_error(rdf_to_rwtbl(keyRdf, scenario = 1:6, add_ym = FALSE))
-  expect_error(rdf_to_rwtbl(keyRdf, scenario = character(0), add_ym = FALSE))
+  expect_warning({
+    expect_error(rdf_to_rwtbl(keyRdf, scenario = c("DNF,CT", "DNF,C1")))
+    expect_error(rdf_to_rwtbl(keyRdf, scenario = 1:6, add_ym = FALSE))
+    expect_error(rdf_to_rwtbl(keyRdf, scenario = character(0), add_ym = FALSE))
+  })
 })
 
 # check the keep_cols option ------------
-rdftbl2 <- rdf_to_rwtbl(keyRdf, keep_cols = TRUE)
-rdftbl3 <- rdf_to_rwtbl(
+rdftbl2 <- expect_warning(rdf_to_rwtbl(keyRdf, keep_cols = TRUE))
+rdftbl3 <- expect_warning(rdf_to_rwtbl(
   keyRdf, 
   keep_cols = c("ObjectName", "Unit"), 
   scenario = 1
-)
+))
 test_that("dimensions are as expected with keep_cols", {
   expect_true(nrow(rdftbl2) > 0)
   expect_equal(ncol(rdftbl2), length(reqCols) + 8)
@@ -136,15 +140,17 @@ test_that("keep_cols warnings post correctly", {
   ) 
   expect_true(all(colnames(tmp) %in% c(reqCols, "Year", "Month", "ObjectName")))
   
-  expect_error(rdf_to_rwtbl(keyRdf, keep_cols = c(FALSE, TRUE)))
-  expect_error(rdf_to_rwtbl(keyRdf, keep_cols = NULL))
+  expect_warning({
+    expect_error(rdf_to_rwtbl(keyRdf, keep_cols = c(FALSE, TRUE)))
+    expect_error(rdf_to_rwtbl(keyRdf, keep_cols = NULL))
+  })
 })
 
 
 # check the default call annual values-----------------
 context("check rdf_to_rwtbl with annual rdf file")
 
-rdftbl <- rdf_to_rwtbl(sysRdf)
+rdftbl <- expect_warning(rdf_to_rwtbl(sysRdf))
 reqCols <- RWDataPlyr:::req_rwtbl_cols()
 nt <- as.integer(sysRdf$meta$number_of_runs)
 slots <- rdf_slot_names(sysRdf)